Bust up unocss theme to make it a bit more readable
