+++ web/src/theme/index.ts
@@ -1,0 +1,3 @@
+export { default as theme, type Theme } from "./theme";
+export { default as rules } from "./rules";
+export { default as shortcuts } from "./shortcuts";
